fix(dev-seed): cast day offset to int for recruiter_daily_activity date
Inserting into the date column made Postgres infer the bound param as a date,
so `current_date - $2` became date-minus-date (integer) and failed. Cast the
offset to int explicitly. Verified against a real Neon dev DB.

+++ backend/scripts/seed-mock-data.mjs
@@ -419,7 +419,7 @@ try {
       await client.query(
         `INSERT INTO public.recruiter_daily_activity
            (recruiter_id, activity_date, submissions_count, interviews_scheduled, interviews_completed, vendor_calls, offers, placements)
-         VALUES ($1, current_date - $2, $3, $4, $5, $6, $7, $8)
+         VALUES ($1, current_date - ($2)::int, $3, $4, $5, $6, $7, $8)
          ON CONFLICT (recruiter_id, activity_date) DO UPDATE SET
            submissions_count = EXCLUDED.submissions_count,
            interviews_scheduled = EXCLUDED.interviews_scheduled,
